feat: 필요한 모델 타입 구현 - 이미지 타입 구현 - ProductDetail, ProductDetailList 중첩 타입 분리 - ProductList 타입 구현
